backend/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.32
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.3
websocket-client==1.8.0
Werkzeug==3.0.3
